dj/crappylife/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
